test(span-login): split huge file tests in smaller ones
