+++ frontend/src/container/SpanDetailsDrawer/constants.ts
@@ -9,3 +9,9 @@ export const RELATED_SIGNALS_VIEW_TYPES = {
 	// METRICS: RelatedSignalsViews.METRICS,
 	INFRA: RelatedSignalsViews.INFRA,
 };
+
+/**
+ * Delay in milliseconds before fetching span percentile data on initial load.
+ * Product requirement to avoid overwhelming API on rapid span selections.
+ */
+export const SPAN_PERCENTILE_INITIAL_DELAY_MS = 2000;
